vce/requirements.txt
asgiref==3.8.1
boto3==1.37.3
botocore==1.37.3
Django==5.1.6
django-storages==1.14.5
djangorestframework==3.16.0
jmespath==1.0.1
mpmath==1.3.0
pillow==11.1.0
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.3
six==1.17.0
sqlparse==0.5.3
sympy==1.13.3
tzdata==2025.1
urllib3==2.3.0
django-cors-headers==4.7.0
django-rest-knox==5.0.2